Move Data Provider button to org page
This button no longer belongs on the pseudo-org page. It belongs on the
actual org page!

Issue #602 Implement UI changes for org-world proposal detail page
